잘못된 disconnect 수정 / bot test 코드 추가

+++ TestClient/TestClient/Utils/Types.h
@@ -46,7 +46,7 @@ using uint16 = unsigned __int16;
 using uint32 = unsigned __int32;
 using uint64 = unsigned __int64;
 
-constexpr uint16 MAX_PACKET_SIZE = 0x10000;
+constexpr int32 MAX_PACKET_SIZE = 0x10000;
 
 #define USING_SHARED_PTR(name)	using name##Ref = std::shared_ptr<class name>;
